fix: dedupe browser cookie handling
