+++ packages/puppeteer-crawler/src/index.ts
@@ -1,6 +1,6 @@
 export * from '@crawlee/browser';
 export * from './internals/puppeteer-crawler';
-export * from './internals/puppeteer-launcher';
+export * from '../puppeteer-launcher';
 
 export * as puppeteerRequestInterception from './internals/utils/puppeteer_request_interception';
 export type { InterceptHandler } from './internals/utils/puppeteer_request_interception';
